Add --clear to venv create.
This fixes a fatal error that would come with a invalid
python symlink in an existing shared library venv.
This is necessary for reinstall-all to be used successfully
to upgrade a system python (Issue #146)

Since we are creating the shared libraries over again,
it is also just cleaner to allow things to start from
scratch.