Remove minLength validation
The min length of any string is 1. Therefore, for consistency, this validation should either be added to all strings or removed from all strings. I decided to remove it. Otherwise, there's too much code duplication. I kept the validator itself in case we have a min length != 1 in the future.