Minor bug in installation instructions for windows
This change will ensure that the command works on Windows, in both `console.exe` and `Powershell`.

Otherwise, in Powershell, if you run the command as currently stated, you will receive the error:

    pipx: The term 'pipx' is not recognized as a name of a cmdlet, function, script file, or executable program.
    Check the spelling of the name, or if a path was included, verify that the path is correct and try again.

Apologies if there is a source page from which this index.html is generated. I couldn't see it.