Fix printing of unavailable apps and man pages
This fixes something which might be a small bug in the original code, but also
affecting the man pages exposition. Post-installation, pipx prints "These apps
are now globally available", followed by a list of app names. However, if all
of the apps are unavailable (`exposed_binary_names` is empty and
`unavailable_binary_names` is non-empty), "These apps are now globally
available" is not printed, when it probably should. In this case, the list of
unavailable apps is printed regardless and appears somewhat out of context.
The same issue affects printing of exposed and unavailable man pages.