Fix combining of `--editable` and `--force` flag (#1124)
* Fix combining of `--editable` and `--force` flag

The 1.3.0 version of `pipx` (with #933) introduced a but where
the `--force-reinstall` flag caused failure when also
the `--editable` flag was used - because it was adding the
`--force-reinstall` flag after `--editable` one (and `--editable`
flag expects url/path to the Python package to install to follow it.

The change fixes it by adding the `--force-reinstall` flag at the
beginning rather than at the end of arguments to avoid this
kind of problem also in case other flags might have similar problem.

Fixes: #1122

* fixup! Fix combining of `--editable` and `--force` flag