Add `--upgrade` argument for install command
When installing a package, both `pip install` and `pipx install` allow
a new package to be installed. If a target package is already installed,
no changes are made. Both pip and pipx also support a forced reinstall
using `--force-reinstall` and `--force` respectively. Another popular
installation more for pip is using the `--upgrade` option, which will
only perform a package change if a new version is detected. Such a
request is not possible in pipx.

This commit adds support for an upgrade method by adding the `--upgrade`
option and forwarding it to pip when a package has been detected as
already installed.

Signed-off-by: James Knight <james.d.knight@live.com>