fix(commands/common/_copy_package_resource): if destination exists and files are same, do not overwrite existing files (#1528)
* fix(commands/common/_copy_package_resource): if destination exists and files are same, do not overwrite existing files

This needs to be performed because on Windows, can_symlink traditionally fails. This means that every time there's an upgrade command run, the previous file is deleted and a new file is created which is a bit wasteful. This ensures that such issues do not happen.

Closes #683

* docs(changelog.d/603.bugfix.md): add

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>