Fix `/changelog/` and `/contributing/` docs URLs
PR #1474 moved the docs to the `docs/` directory, and also changed
casing of some filenames.

- `docs/changelog.md` -> `docs/CHANGELOG.md`
- `docs/contributing.md` -> `docs/CONTRIBUTING.md`

This changed URL path parameters as well. URL paths to these pages must
now be uppercased when accessing the deployed docs (#1540).

For example, navigating to https://pipx.pypa.io/stable/CONTRIBUTING/
shows the correct page, but navigating to
https://pipx.pypa.io/stable/contributing/ displays a (poorly styled)
`404` page.

This commit will move the pages back to `docs/changelog.md` and
`docs/contributing.md` to restore the previous URLs to these pages.

Note that the casing changes to not affect the homepage because MkDocs
automatically handles `README.md` as `index.md`.
https://www.mkdocs.org/user-guide/writing-your-docs/#index-pages